Refactoring with ESlint, rename containers => components & components => pages
