+++ src/components/ProductInfo.js
